Install packages required by camtrapReport

Install packages required for the full camtrapReport workflow, including packages used by optional report modules.

Usage

install_All(pkgs = NULL, update = FALSE, ...)

Arguments

pkgs

An optional character vector of additional CRAN package names. The default is NULL.

update

A logical value. If TRUE, optional packages are reinstalled. The default is FALSE.

...

Additional arguments passed to install.packages().

Value

Called for its side effects. The function returns NULL invisibly.

Details

The function checks the package list used by camtrapReport and installs packages that are not currently available. Packages listed in the package configuration as GitHub dependencies are installed with remotes::install_github().

See also

Examples

if (interactive()) {
  # These calls install packages into the user's R library.
  install_All()
  install_All(pkgs = "remotes")
  install_All(update = TRUE)
}
#> All required packages have already been installed.
#> All required packages have already been installed.
#> Installing package into ‘C:/Users/ebrah010/AppData/Local/R/win-library/4.5’
#> (as ‘lib’ is unspecified)
#> package ‘gt’ successfully unpacked and MD5 sums checked
#> 
#> The downloaded binary packages are in
#> 	C:\Users\ebrah010\AppData\Local\Temp\RtmpS261EZ\downloaded_packages
#> Warning: package ‘dplyr’ is in use and will not be installed
#> Installing package into ‘C:/Users/ebrah010/AppData/Local/R/win-library/4.5’
#> (as ‘lib’ is unspecified)
#> package ‘leaflet’ successfully unpacked and MD5 sums checked
#> 
#> The downloaded binary packages are in
#> 	C:\Users\ebrah010\AppData\Local\Temp\RtmpS261EZ\downloaded_packages
#> Installing package into ‘C:/Users/ebrah010/AppData/Local/R/win-library/4.5’
#> (as ‘lib’ is unspecified)
#> package ‘terra’ successfully unpacked and MD5 sums checked
#> Warning: cannot remove prior installation of package ‘terra’
#> Warning: problem copying C:\Users\ebrah010\AppData\Local\R\win-library\4.5\00LOCK\terra\libs\x64\terra.dll to C:\Users\ebrah010\AppData\Local\R\win-library\4.5\terra\libs\x64\terra.dll: Permission denied
#> Warning: restored ‘terra’
#> 
#> The downloaded binary packages are in
#> 	C:\Users\ebrah010\AppData\Local\Temp\RtmpS261EZ\downloaded_packages
#> Installing package into ‘C:/Users/ebrah010/AppData/Local/R/win-library/4.5’
#> (as ‘lib’ is unspecified)
#> package ‘htmltools’ successfully unpacked and MD5 sums checked
#> Warning: cannot remove prior installation of package ‘htmltools’
#> Warning: problem copying C:\Users\ebrah010\AppData\Local\R\win-library\4.5\00LOCK\htmltools\libs\x64\htmltools.dll to C:\Users\ebrah010\AppData\Local\R\win-library\4.5\htmltools\libs\x64\htmltools.dll: Permission denied
#> Warning: restored ‘htmltools’
#> 
#> The downloaded binary packages are in
#> 	C:\Users\ebrah010\AppData\Local\Temp\RtmpS261EZ\downloaded_packages
#> Installing package into ‘C:/Users/ebrah010/AppData/Local/R/win-library/4.5’
#> (as ‘lib’ is unspecified)
#> package ‘dygraphs’ successfully unpacked and MD5 sums checked
#> 
#> The downloaded binary packages are in
#> 	C:\Users\ebrah010\AppData\Local\Temp\RtmpS261EZ\downloaded_packages
#> Installing package into ‘C:/Users/ebrah010/AppData/Local/R/win-library/4.5’
#> (as ‘lib’ is unspecified)
#> package ‘xts’ successfully unpacked and MD5 sums checked
#> 
#> The downloaded binary packages are in
#> 	C:\Users\ebrah010\AppData\Local\Temp\RtmpS261EZ\downloaded_packages
#> Installing package into ‘C:/Users/ebrah010/AppData/Local/R/win-library/4.5’
#> (as ‘lib’ is unspecified)
#> package ‘htmlwidgets’ successfully unpacked and MD5 sums checked
#> 
#> The downloaded binary packages are in
#> 	C:\Users\ebrah010\AppData\Local\Temp\RtmpS261EZ\downloaded_packages
#> Warning: package ‘ggplot2’ is in use and will not be installed
#> Installing package into ‘C:/Users/ebrah010/AppData/Local/R/win-library/4.5’
#> (as ‘lib’ is unspecified)
#> package ‘ggrepel’ successfully unpacked and MD5 sums checked
#> 
#> The downloaded binary packages are in
#> 	C:\Users\ebrah010\AppData\Local\Temp\RtmpS261EZ\downloaded_packages
#> Installing package into ‘C:/Users/ebrah010/AppData/Local/R/win-library/4.5’
#> (as ‘lib’ is unspecified)
#> package ‘activity’ successfully unpacked and MD5 sums checked
#> 
#> The downloaded binary packages are in
#> 	C:\Users\ebrah010\AppData\Local\Temp\RtmpS261EZ\downloaded_packages
#> Installing package into ‘C:/Users/ebrah010/AppData/Local/R/win-library/4.5’
#> (as ‘lib’ is unspecified)
#> package ‘corrplot’ successfully unpacked and MD5 sums checked
#> 
#> The downloaded binary packages are in
#> 	C:\Users\ebrah010\AppData\Local\Temp\RtmpS261EZ\downloaded_packages
#> Installing package into ‘C:/Users/ebrah010/AppData/Local/R/win-library/4.5’
#> (as ‘lib’ is unspecified)
#> package ‘spatstat’ successfully unpacked and MD5 sums checked
#> 
#> The downloaded binary packages are in
#> 	C:\Users\ebrah010\AppData\Local\Temp\RtmpS261EZ\downloaded_packages
#> Installing package into ‘C:/Users/ebrah010/AppData/Local/R/win-library/4.5’
#> (as ‘lib’ is unspecified)
#> package ‘scales’ successfully unpacked and MD5 sums checked
#> 
#> The downloaded binary packages are in
#> 	C:\Users\ebrah010\AppData\Local\Temp\RtmpS261EZ\downloaded_packages
#> Installing package into ‘C:/Users/ebrah010/AppData/Local/R/win-library/4.5’
#> (as ‘lib’ is unspecified)
#> package ‘iNEXT’ successfully unpacked and MD5 sums checked
#> 
#> The downloaded binary packages are in
#> 	C:\Users\ebrah010\AppData\Local\Temp\RtmpS261EZ\downloaded_packages
#> Installing package into ‘C:/Users/ebrah010/AppData/Local/R/win-library/4.5’
#> (as ‘lib’ is unspecified)
#> package ‘tidyr’ successfully unpacked and MD5 sums checked
#> 
#> The downloaded binary packages are in
#> 	C:\Users\ebrah010\AppData\Local\Temp\RtmpS261EZ\downloaded_packages
#> Installing package into ‘C:/Users/ebrah010/AppData/Local/R/win-library/4.5’
#> (as ‘lib’ is unspecified)
#> package ‘magick’ successfully unpacked and MD5 sums checked
#> 
#> The downloaded binary packages are in
#> 	C:\Users\ebrah010\AppData\Local\Temp\RtmpS261EZ\downloaded_packages
#> Installing package into ‘C:/Users/ebrah010/AppData/Local/R/win-library/4.5’
#> (as ‘lib’ is unspecified)
#> package ‘remotes’ successfully unpacked and MD5 sums checked
#> 
#> The downloaded binary packages are in
#> 	C:\Users\ebrah010\AppData\Local\Temp\RtmpS261EZ\downloaded_packages
#> Installing package into ‘C:/Users/ebrah010/AppData/Local/R/win-library/4.5’
#> (as ‘lib’ is unspecified)
#> package ‘suncalc’ successfully unpacked and MD5 sums checked
#> 
#> The downloaded binary packages are in
#> 	C:\Users\ebrah010\AppData\Local\Temp\RtmpS261EZ\downloaded_packages
#> Installing package into ‘C:/Users/ebrah010/AppData/Local/R/win-library/4.5’
#> (as ‘lib’ is unspecified)
#> package ‘taxize’ successfully unpacked and MD5 sums checked
#> 
#> The downloaded binary packages are in
#> 	C:\Users\ebrah010\AppData\Local\Temp\RtmpS261EZ\downloaded_packages
#> Installing package into ‘C:/Users/ebrah010/AppData/Local/R/win-library/4.5’
#> (as ‘lib’ is unspecified)
#> package ‘lubridate’ successfully unpacked and MD5 sums checked
#> Warning: cannot remove prior installation of package ‘lubridate’
#> Warning: problem copying C:\Users\ebrah010\AppData\Local\R\win-library\4.5\00LOCK\lubridate\libs\x64\lubridate.dll to C:\Users\ebrah010\AppData\Local\R\win-library\4.5\lubridate\libs\x64\lubridate.dll: Permission denied
#> Warning: restored ‘lubridate’
#> 
#> The downloaded binary packages are in
#> 	C:\Users\ebrah010\AppData\Local\Temp\RtmpS261EZ\downloaded_packages
#> Installing package into ‘C:/Users/ebrah010/AppData/Local/R/win-library/4.5’
#> (as ‘lib’ is unspecified)
#> package ‘lutz’ successfully unpacked and MD5 sums checked
#> Warning: cannot remove prior installation of package ‘lutz’
#> Warning: problem copying C:\Users\ebrah010\AppData\Local\R\win-library\4.5\00LOCK\lutz\libs\x64\lutz.dll to C:\Users\ebrah010\AppData\Local\R\win-library\4.5\lutz\libs\x64\lutz.dll: Permission denied
#> Warning: restored ‘lutz’
#> 
#> The downloaded binary packages are in
#> 	C:\Users\ebrah010\AppData\Local\Temp\RtmpS261EZ\downloaded_packages
#> Installing package into ‘C:/Users/ebrah010/AppData/Local/R/win-library/4.5’
#> (as ‘lib’ is unspecified)
#> package ‘sf’ successfully unpacked and MD5 sums checked
#> Warning: cannot remove prior installation of package ‘sf’
#> Warning: problem copying C:\Users\ebrah010\AppData\Local\R\win-library\4.5\00LOCK\sf\libs\x64\sf.dll to C:\Users\ebrah010\AppData\Local\R\win-library\4.5\sf\libs\x64\sf.dll: Permission denied
#> Warning: restored ‘sf’
#> 
#> The downloaded binary packages are in
#> 	C:\Users\ebrah010\AppData\Local\Temp\RtmpS261EZ\downloaded_packages
#> Installing package into ‘C:/Users/ebrah010/AppData/Local/R/win-library/4.5’
#> (as ‘lib’ is unspecified)
#> package ‘sbd’ successfully unpacked and MD5 sums checked
#> 
#> The downloaded binary packages are in
#> 	C:\Users\ebrah010\AppData\Local\Temp\RtmpS261EZ\downloaded_packages
#> Installing package into ‘C:/Users/ebrah010/AppData/Local/R/win-library/4.5’
#> (as ‘lib’ is unspecified)
#> package ‘Distance’ successfully unpacked and MD5 sums checked
#> 
#> The downloaded binary packages are in
#> 	C:\Users\ebrah010\AppData\Local\Temp\RtmpS261EZ\downloaded_packages
#> Warning: package ‘readr’ is in use and will not be installed
#> 27 packages were successfully reinstalled.
#> All requested optional packages are installed.