camReport classes

Class definitions used by camtrapReport to store camera-trap data, report metadata, report sections, and intermediate report objects.

Value

A camReport reference-class object containing the imported camera-trap data, metadata, settings, analysis results, and report components.

camReport class

The camReport class is the main object class used by camtrapReport. Objects of this class are usually created with camData(). A camReport object stores the camera-trap dataset, report metadata, processed summaries, report modules, status-report modules, and settings used for report generation.

Main fields

Important fields in a camReport object include:

data: A list containing the camera-trap data tables, including observations, deployments, media, locations, sequences, and taxonomy.

habitat: Optional habitat information linked to camera locations.

study_area: Optional spatial object or path describing the study area.

siteName: Name of the study site.

title: Title used in the generated report.

subtitle: Subtitle used in the generated report.

authors: Author or contributor text used in the generated report.

institute: Institute or organisation text used in the generated report.

description: Study-area or project description used in the generated report.

years: Years included in the report summaries.

group_definition: Definitions of species or observation groups used in the report.

setting: Report settings, including selected focus groups and other configuration options.

data_status: Data-status summaries generated from the input dataset.

reportObjectElements: Report modules and related objects used to generate the ecological report.

statusReportObjects: Objects used to generate the data-status report.

Supporting classes

camtrapReport also defines supporting classes and class unions used internally, including camInfo, .Rchunk, .textSection, characterORnull, characterORlist, characterORlistORnull, listORnull, and data.frameORnull.

See also

Examples

# Inspect the formal definition of the main report class.
methods::getClass("camReport")
#> Reference Class "camReport":
#> 
#> Class fields:
#>                                                             
#> Name:                        data                    habitat
#> Class:                       list                 data.frame
#>                                                             
#> Name:                 data_merged                 study_area
#> Class:                 data.frame                 listORnull
#>                                                             
#> Name:                    siteName                      title
#> Class:                  character                  character
#>                                                             
#> Name:                    subtitle                    authors
#> Class:                  character                  character
#>                                                             
#> Name:                   institute                   logoPath
#> Class:                  character                  character
#>                                                             
#> Name:                   annonator            acknowledgement
#> Class:                  character                  character
#>                                                             
#> Name:                 description                      years
#> Class:                  character                    numeric
#>                                                             
#> Name:            group_definition              filterExclude
#> Class:                       list                       list
#>                                                             
#> Name:                  filterKeep                filterCount
#> Class:                       list                    numeric
#>                                                             
#> Name:              filterDuration            observed_counts
#> Class:                    numeric                 data.frame
#>                                                             
#> Name:            frequent_species               camera_stats
#> Class:                 data.frame                 data.frame
#>                                                             
#> Name:                camera_setup          observation_stats
#> Class:                 data.frame                 data.frame
#>                                                             
#> Name:             species_summary              species_stats
#> Class:                       list                 data.frame
#>                                                             
#> Name:  species_summary_by_habitat                    capture
#> Class:                 data.frame                 data.frame
#>                                                             
#> Name:           density_estimates                    setting
#> Class:                 data.frame                       list
#>                                                             
#> Name:                   sun_times                   packages
#> Class:           data.frameORnull                  character
#>                                                             
#> Name:                        info                data_status
#> Class:                       list                       list
#>                                                             
#> Name:                         rem                .rem_params
#> Class:                       list                       list
#>                                                             
#> Name:                 .act_models          .any_data_for_rem
#> Class:                       list                    logical
#>                                                             
#> Name:                .tempObjects         reportTextElements
#> Class:                       list                       list
#>                                                             
#> Name:        reportObjectElements        statusReportObjects
#> Class:                       list                       list
#>                                  
#> Name:               reportObjects
#> Class:                       list
#> 
#> Class Methods: 
#>      "gui", "generateStatusReport", "generateReport", "extractYears", 
#>      "get_data_subset", "addStatusReportObject", "addReportObject", "filter", 
#>      "show#envRefClass", "setup", "get_REM", "fit_REM", ".get_REM_Param", 
#>      "species_summary_by_location", "spatial_density", "richness", 
#>      "recetFigTabNumber", "getTableNumber", "getFigureNumber", 
#>      "get_speciesNames", "get_focus_group", "set_focus_group", "get_group", 
#>      "add_group", "initialize", "field", "trace", "getRefClass", "initFields", 
#>      "copy", "callSuper", ".objectPackage", "export", "untrace", "getClass", 
#>      "show", "usingMethods", ".objectParent", "import"
#> 
#> Reference Superclasses: 
#>      "envRefClass"
#> 

# Check whether the class is registered.
methods::isClass("camReport")
#> [1] TRUE